src/FunctionsMcpTool/requirements.txt
# Do not include azure-functions-worker in this file
azure-functions
azure-identity
requests
python-tds
azure-functions-durable
botbuilder-core
botbuilder-integration-aiohttp
azure-ai-projects
